@@ -1,53 +1,34 @@
 {
   "_args": [
     [
-      {
-        "raw": "ios-sim@^5.0.12",
-        "scope": null,
-        "escapedName": "ios-sim",
-        "name": "ios-sim",
-        "rawSpec": "^5.0.12",
-        "spec": ">=5.0.12 <6.0.0",
-        "type": "range"
-      },
-      "/Users/shazron/Documents/git/apache/cordova-ios"
+      "ios-sim@6.1.2",
+      "/Users/spindori/Documents/Cordova/cordova-ios"
     ]
   ],
-  "_from": "ios-sim@>=5.0.12 <6.0.0",
-  "_id": "ios-sim@5.0.12",
+  "_from": "ios-sim@6.1.2",
+  "_id": "ios-sim@6.1.2",
   "_inBundle": true,
-  "_inCache": true,
+  "_integrity": "sha512-kWSc7XJfYAGaQSh51UM4pDgXc77vqBS62oBA2tnXsHajk2+xl7Oyt613GYq64eMry2BM1F3tJLGFCJfpp1Ge+A==",
   "_location": "/cordova-ios/ios-sim",
-  "_nodeVersion": "6.7.0",
-  "_npmOperationalInternal": {
-    "host": "packages-12-west.internal.npmjs.com",
-    "tmp": "tmp/ios-sim-5.0.12.tgz_1480357903209_0.8315212763845921"
-  },
-  "_npmUser": {
-    "name": "shazron",
-    "email": "shazron@gmail.com"
-  },
-  "_npmVersion": "3.10.9",
   "_phantomChildren": {
-    "abbrev": "1.0.9"
+    "abbrev": "1.1.1"
   },
   "_requested": {
-    "raw": "ios-sim@^5.0.12",
-    "scope": null,
-    "escapedName": "ios-sim",
+    "type": "version",
+    "registry": true,
+    "raw": "ios-sim@6.1.2",
     "name": "ios-sim",
-    "rawSpec": "^5.0.12",
-    "spec": ">=5.0.12 <6.0.0",
-    "type": "range"
+    "escapedName": "ios-sim",
+    "rawSpec": "6.1.2",
+    "saveSpec": null,
+    "fetchSpec": "6.1.2"
   },
   "_requiredBy": [
     "/cordova-ios"
   ],
-  "_resolved": "https://registry.npmjs.org/ios-sim/-/ios-sim-5.0.12.tgz",
-  "_shasum": "0b07436ce1bf663bda9a3fb75412f41a1dec465b",
-  "_shrinkwrap": null,
-  "_spec": "ios-sim@^5.0.12",
-  "_where": "/Users/shazron/Documents/git/apache/cordova-ios",
+  "_resolved": "https://registry.npmjs.org/ios-sim/-/ios-sim-6.1.2.tgz",
+  "_spec": "6.1.2",
+  "_where": "/Users/spindori/Documents/Cordova/cordova-ios",
   "author": {
     "name": "Shazron Abdullah"
   },
@@ -61,23 +42,17 @@
     "bplist-parser": "^0.0.6",
     "nopt": "1.0.9",
     "plist": "^1.2.0",
-    "simctl": "^0.1.0"
+    "simctl": "^1.1.1"
   },
-  "description": "launch iOS apps into the iOS Simulator from the command line (Xcode 7.0+)",
+  "description": "launch iOS apps into the iOS Simulator from the command line (Xcode 8.0+)",
   "devDependencies": {
-    "jasmine-node": "^1.14.5",
+    "jasmine": "~2.6.0",
     "jscs": "^2.11.0",
     "jshint": "^2.9.1"
   },
-  "directories": {},
-  "dist": {
-    "shasum": "0b07436ce1bf663bda9a3fb75412f41a1dec465b",
-    "tarball": "https://registry.npmjs.org/ios-sim/-/ios-sim-5.0.12.tgz"
-  },
   "engines": {
-    "node": ">=0.10.0"
+    "node": ">=4"
   },
-  "gitHead": "0e0cafacca40434702a53538b11169a345070495",
   "homepage": "https://github.com/phonegap/ios-sim#readme",
   "keywords": [
     "ios-sim",
@@ -85,38 +60,19 @@
   ],
   "license": "MIT",
   "main": "ios-sim.js",
-  "maintainers": [
-    {
-      "name": "macdonst",
-      "email": "simon.macdonald@gmail.com"
-    },
-    {
-      "name": "purplecabbage",
-      "email": "purplecabbage@gmail.com"
-    },
-    {
-      "name": "shazron",
-      "email": "shazron@gmail.com"
-    },
-    {
-      "name": "stevegill",
-      "email": "stevengill97@gmail.com"
-    }
-  ],
   "name": "ios-sim",
-  "optionalDependencies": {},
   "preferGlobal": "true",
   "repository": {
     "type": "git",
     "url": "git+https://github.com/phonegap/ios-sim.git"
   },
   "scripts": {
-    "jasmine": "jasmine-node --captureExceptions --color spec",
+    "jasmine": "jasmine --config=spec/jasmine.json",
     "jscs": "jscs src ./ios-sim.js",
     "jshint": "jshint src ./ios-sim.js",
     "postjshint": "npm run jscs",
     "posttest": "npm run jshint",
     "test": "npm run jasmine"
   },
-  "version": "5.0.12"
+  "version": "6.1.2"
 }
