@@ -1,46 +1,32 @@
 {
   "_args": [
     [
-      {
-        "raw": "stream-buffers@~0.2.3",
-        "scope": null,
-        "escapedName": "stream-buffers",
-        "name": "stream-buffers",
-        "rawSpec": "~0.2.3",
-        "spec": ">=0.2.3 <0.3.0",
-        "type": "range"
-      },
-      "/Users/shazron/Documents/git/apache/cordova-ios/node_modules/bplist-creator"
+      "stream-buffers@2.2.0",
+      "/Users/spindori/Documents/Cordova/cordova-ios"
     ]
   ],
-  "_from": "stream-buffers@>=0.2.3 <0.3.0",
-  "_id": "stream-buffers@0.2.6",
+  "_from": "stream-buffers@2.2.0",
+  "_id": "stream-buffers@2.2.0",
   "_inBundle": true,
-  "_inCache": true,
+  "_integrity": "sha1-kdX1Ew0c75bc+n9yaUUYh0HQnuQ=",
   "_location": "/cordova-ios/stream-buffers",
-  "_npmUser": {
-    "name": "samcday",
-    "email": "me@samcday.com.au"
-  },
-  "_npmVersion": "2.0.0-alpha-5",
   "_phantomChildren": {},
   "_requested": {
-    "raw": "stream-buffers@~0.2.3",
-    "scope": null,
-    "escapedName": "stream-buffers",
+    "type": "version",
+    "registry": true,
+    "raw": "stream-buffers@2.2.0",
     "name": "stream-buffers",
-    "rawSpec": "~0.2.3",
-    "spec": ">=0.2.3 <0.3.0",
-    "type": "range"
+    "escapedName": "stream-buffers",
+    "rawSpec": "2.2.0",
+    "saveSpec": null,
+    "fetchSpec": "2.2.0"
   },
   "_requiredBy": [
     "/cordova-ios/bplist-creator"
   ],
-  "_resolved": "https://registry.npmjs.org/stream-buffers/-/stream-buffers-0.2.6.tgz",
-  "_shasum": "181c08d5bb3690045f69401b9ae6a7a0cf3313fc",
-  "_shrinkwrap": null,
-  "_spec": "stream-buffers@~0.2.3",
-  "_where": "/Users/shazron/Documents/git/apache/cordova-ios/node_modules/bplist-creator",
+  "_resolved": "https://registry.npmjs.org/stream-buffers/-/stream-buffers-2.2.0.tgz",
+  "_spec": "2.2.0",
+  "_where": "/Users/spindori/Documents/Cordova/cordova-ios",
   "author": {
     "name": "Sam Day",
     "email": "me@samcday.com.au"
@@ -51,41 +37,27 @@
   "dependencies": {},
   "description": "Buffer-backed Streams for reading and writing.",
   "devDependencies": {
+    "istanbul": "~0.3.2",
     "vows": ">= 0.5.6"
   },
-  "directories": {},
-  "dist": {
-    "shasum": "181c08d5bb3690045f69401b9ae6a7a0cf3313fc",
-    "tarball": "https://registry.npmjs.org/stream-buffers/-/stream-buffers-0.2.6.tgz"
-  },
   "engines": {
-    "node": ">= 0.3.0"
+    "node": ">= 0.10.0"
   },
-  "gitHead": "b1dd62eb2e127d48ed7f1e814d5451d2b0dec144",
-  "homepage": "https://github.com/samcday/node-stream-buffer",
+  "homepage": "https://github.com/samcday/node-stream-buffer#readme",
   "keywords": [
     "memory streams",
     "streams",
     "buffer streams"
   ],
+  "license": "Unlicense",
   "main": "./lib/streambuffer.js",
-  "maintainers": [
-    {
-      "name": "samcday",
-      "email": "sam.c.day@gmail.com"
-    }
-  ],
   "name": "stream-buffers",
-  "optionalDependencies": {},
-  "publishConfig": {
-    "registry": "https://registry.npmjs.org"
-  },
   "repository": {
     "type": "git",
     "url": "git+https://github.com/samcday/node-stream-buffer.git"
   },
   "scripts": {
-    "test": "vows --spec"
+    "test": "[ -n \"$NO_COVERAGE\" ] && vows --spec || istanbul cover vows -- --spec"
   },
-  "version": "0.2.6"
+  "version": "2.2.0"
 }
